🔨 use GrapherState to figure out chart type
